--- test/mitmproxy/addons/test_defaults.py
@@ -1,5 +1,0 @@
-from mitmproxy import addons
-
-
-def test_defaults():
-    assert addons.default_addons()
