--- test/mitmproxy/net/test_imports.py
@@ -1,1 +1,0 @@
-# These are actually tests!
